Update RECTANGLE PARTITION
There is a rectangle of given width w and height h, On the width side, you are given a list of measurements. On the height side, you are given another list of measurements.
Draw perpendicular lines from the measurements to partition the rectangle into smaller rectangles. In all sub-rectangles (include the combinations of smaller rectangles), how many of them are squares?
Input: 
10 5 2 1
2 5
3
Output
4